ab_core_primitives::block::header

Module owned

Data structures related to the owned version of BlockHeader

Structs

OwnedBeaconChainHeader

An owned version of BeaconChainHeader.

OwnedBeaconChainHeaderUnsealed

Owned beacon chain block header, which is not sealed yet

OwnedIntermediateShardHeader

An owned version of IntermediateShardHeader.

OwnedIntermediateShardHeaderUnsealed

Owned intermediate shard block header, which is not sealed yet

OwnedLeafShardHeader

An owned version of LeafShardHeader.

OwnedLeafShardHeaderUnsealed

Owned leaf shard block header, which is not sealed yet

Enums

OwnedBeaconChainHeaderError

Errors for OwnedBeaconChainHeader

OwnedBlockHeader

An owned version of BlockHeader.

OwnedIntermediateShardHeaderError

Errors for OwnedIntermediateShardHeader

Traits

GenericOwnedBlockHeader

Generic owned block header