ab_core_primitives::block::header

Struct BlockHeaderConsensusParameters

pub struct BlockHeaderConsensusParameters<'a> {
    pub fixed_parameters: BlockHeaderFixedConsensusParameters,
    pub super_segment_root: Option<&'a SuperSegmentRoot>,
    pub next_solution_range: Option<SolutionRange>,
    pub pot_parameters_change: Option<&'a BlockHeaderPotParametersChange>,
}

Consensus parameters (on the beacon chain)

Fields
fixed_parameters: BlockHeaderFixedConsensusParameters

Consensus parameters that are always present

super_segment_root: Option<&'a SuperSegmentRoot>

Super segment root

next_solution_range: Option<SolutionRange>

Solution range for the next block/interval (if any)

pot_parameters_change: Option<&'a BlockHeaderPotParametersChange>

Change of parameters to apply to the proof of time chain (if any)

Implementations

impl<'a> BlockHeaderConsensusParameters<'a>

pub const MAX_SIZE: u32

Max size of the allocation necessary for this data structure

pub const SUPER_SEGMENT_ROOT_MASK: u8 = 0b_0000_0001

Bitmask for presence of super_segment_root field

pub const NEXT_SOLUTION_RANGE_MASK: u8 = 0b_0000_0010

Bitmask for presence of next_solution_range field

pub const POT_PARAMETERS_CHANGE_MASK: u8 = 0b_0000_0100

Bitmask for presence of pot_parameters_change field

pub const MASK_ALL: u8

All supported bitmask variants

pub fn try_from_bytes(bytes: &'a [u8]) -> Option<(Self, &'a [u8])>

Create an instance from provided bytes.

bytes do not need to be aligned.

Returns an instance and remaining bytes on success.

pub fn hash(&self) -> Blake3Hash

Hash of the block consensus parameters, part of the eventual block root