ab_core_primitives::block::header

Struct IntermediateShardHeader

#[non_exhaustive]
pub struct IntermediateShardHeader<'a> { /* private fields */ }

Block header that corresponds to an intermediate shard

Implementations

impl<'a> IntermediateShardHeader<'a>

pub fn try_from_bytes(bytes: &'a [u8]) -> Option<(Self, &'a [u8])>

Try to create a new instance from provided bytes.

bytes should be 8-bytes aligned.

Returns an instance and remaining bytes on success, None if too few bytes were given, bytes are not properly aligned or input is otherwise invalid.

pub fn is_internally_consistent(&self) -> bool

Check block headerโs internal consistency.

This is usually not necessary to be called explicitly since internal consistency is checked by Self::try_from_bytes() internally.

pub fn try_from_bytes_unchecked(bytes: &'a [u8]) -> Option<(Self, &'a [u8])>

The same as Self::try_from_bytes(), but for trusted input that skips some consistency checks

pub fn to_owned(self) -> OwnedIntermediateShardHeader

Create an owned version of this header

pub fn shared(&self) -> &SharedBlockHeader<'a>

Shared block header

pub fn beacon_chain_info(&self) -> &'a BlockHeaderBeaconChainInfo

Beacon chain info

pub fn child_shard_blocks(&self) -> &BlockHeaderChildShardBlocks<'a>

Information about child shard blocks

pub fn pre_seal_hash(&self) -> Blake3Hash

Hash of the block before seal is applied to it

pub fn is_sealed_correctly(&self) -> bool

Verify seal against IntermediateShardHeader::pre_seal_hash() and check that its public key hash corresponds to the solution

pub fn root(&self) -> impl Deref<Target = BlockRoot> + Send + Sync

Compute block root out of this header.

Block root is a Merkle Tree Root. The leaves are derived from individual fields in SharedBlockHeader and other fields of this enum in the declaration order.

Note that this method computes root by doing a bunch of hashing. The result is then cached if alloc feature is enabled or when compiled for OS target that is not none.