ab_core_primitives::block

Module owned

Data structures related to the owned version of Block

Structs

OwnedBeaconChainBlock

An owned version of BeaconChainBlock.

OwnedBeaconChainBlockBuilder

Builder for OwnedBeaconChainBlock

OwnedBeaconChainBlockUnsealed

Owned beacon chain block header, which is not sealed yet

OwnedIntermediateShardBlock

An owned version of IntermediateShardBlock.

OwnedIntermediateShardBlockBuilder

Builder for OwnedIntermediateShardBlock

OwnedIntermediateShardBlockUnsealed

Owned intermediate shard block header, which is not sealed yet

OwnedLeafShardBlock

An owned version of LeafShardBlock.

OwnedLeafShardBlockBuilder

Builder for OwnedLeafShardBlock

OwnedLeafShardBlockUnsealed

Owned leaf shard block header, which is not sealed yet

Enums

OwnedBlock

An owned version of Block.

OwnedIntermediateShardBlockError

Errors for OwnedIntermediateShardBlock

Traits

GenericOwnedBlock

Generic owned block