ab_core_primitives

Module segments

Segments-related data structures

Structs

ArchivedBlockProgress

Progress of an archived block.

ArchivedHistorySegment

Archived history segment after archiving is applied.

HistorySize

Size of blockchain history in segments

LastArchivedBlock

Last archived block

LocalSegmentIndex

Local segment index of a shard

RecordedHistorySegment

Recorded history segment before archiving is applied.

SegmentHeader

Segment header for a specific segment of a shard

SegmentIndex

Segment index

SegmentPosition

Segment position in a super segment

SegmentRoot

Segment root contained within a segment

ShardSegmentRootWithPosition

Shard segment root with position

SuperSegment

Super segment

SuperSegmentHeader

Super segment header

SuperSegmentIndex

Super segment index

SuperSegmentRoot

Super segment root contained within a beacon chain block