ab_core_primitives::segments

Struct ArchivedBlockProgress

pub struct ArchivedBlockProgress { /* private fields */ }

Progress of an archived block.

Implementations

impl ArchivedBlockProgress

pub const fn new_complete() -> Self

Block is archived fully

pub const fn new_partial(new_partial: NonZeroU32) -> Self

Block is partially archived with the provided number of bytes

pub const fn partial(&self) -> Option<NonZeroU32>

Return the number of partially archived bytes if the progress is not complete