ab_core_primitives::solutions

Struct Solution

#[repr(C)]
pub struct Solution {
    pub public_key_hash: Blake3Hash,
    pub shard_commitment: SolutionShardCommitment,
    pub record_root: RecordRoot,
    pub record_proof: RecordProof,
    pub chunk: RecordChunk,
    pub chunk_proof: ChunkProof,
    pub proof_of_space: PosProof,
    pub history_size: HistorySize,
    pub sector_index: SectorIndex,
    pub piece_offset: PieceOffset,
    pub padding: [u8; 4],
}

Farmer solution for slot challenge.

Fields
public_key_hash: Blake3Hash

Public key of the farmer that created the solution

shard_commitment: SolutionShardCommitment

Farmer’s shard commitment

record_root: RecordRoot

Record root that can use used to verify that piece was included in blockchain history

record_proof: RecordProof

Proof for above record root

chunk: RecordChunk

Chunk at the above offset

chunk_proof: ChunkProof

Proof for the above chunk

proof_of_space: PosProof

Proof of space for piece offset

history_size: HistorySize

Size of the blockchain history at the time of sector creation

sector_index: SectorIndex

Index of the sector where the solution was found

piece_offset: PieceOffset

Pieces offset within sector

padding: [u8; 4]

Padding for data structure alignment

Implementations

impl Solution

pub fn genesis_solution() -> Self

Fake solution for the genesis block

pub fn verify<PotVerifier>( &self, slot: SlotNumber, params: &SolutionVerifyParams, ) -> Result<(), SolutionVerifyError>
where PotVerifier: SolutionPotVerifier,

Check solution validity