ab_core_primitives

Module transaction

Transaction-related primitives

Modules

owned

Data structures related to the owned version of Transaction

Structs

Gas

A measure of compute resources, 1 Gas == 1 ns of compute on reference hardware

SerializedTransactionLengths

Lengths of various components in a serialized version of Transaction

Transaction

Similar to Transaction, but doesn’t require allow or data ownership.

TransactionHash

Transaction hash

TransactionHeader

Transaction header

TransactionSlot

Transaction slot