ab_core_primitives::block

Module body

Block body primitives

Modules

owned

Data structures related to the owned version of BlockBody

Structs

BeaconChainBody

Block body that corresponds to the beacon chain

IntermediateShardBlockInfo

Information about intermediate shard block

IntermediateShardBlocksInfo

Information about a collection of intermediate shard blocks

IntermediateShardBody

Block body that corresponds to an intermediate shard

LeafShardBlockInfo

Information about leaf shard block container inside intermediate shard block body

LeafShardBlocksInfo

Information about a collection of leaf shard blocks

LeafShardBody

Block body that corresponds to a leaf shard

LeafShardOwnSegments

Information about leaf shard segments

OwnSegments

Own segments produced by a shard

Transactions

Collection of transactions

Enums

BlockBody

Block body that together with BlockHeader form a Block

Traits

GenericBlockBody

Generic block body

Functions

compute_segments_root

Calculates a Merkle Tree root for a provided list of segment roots