ab_core_primitives::block

Module header

Block header primitives

Modules

owned

Data structures related to the owned version of BlockHeader

Structs

BeaconChainHeader

Block header that corresponds to the beacon chain

BlockHeaderBeaconChainInfo

Beacon chain info

BlockHeaderChildShardBlocks

Information about child shard blocks

BlockHeaderConsensusInfo

Consensus information in the block header

BlockHeaderConsensusParameters

Consensus parameters (on the beacon chain)

BlockHeaderEd25519Seal

Ed25519 seal

BlockHeaderFixedConsensusParameters

Consensus parameters (on the beacon chain)

BlockHeaderPotParametersChange

A mirror of PotParametersChange for block header purposes.

BlockHeaderPrefix

Block header prefix.

BlockHeaderResult

Block header result.

IntermediateShardHeader

Block header that corresponds to an intermediate shard

LeafShardHeader

Block header that corresponds to a leaf shard

OwnedBlockHeaderConsensusParameters

Owned version of BlockHeaderConsensusParameters

SharedBlockHeader

Part of the block header, shared for different kinds of shards

Enums

BlockHeader

Block header that together with BlockBody form a Block

BlockHeaderSeal

Block header seal

BlockHeaderSealType

Block header seal type

OwnedBlockHeaderSeal

Owned version of BlockHeaderSeal

Traits

GenericBlockHeader

Generic block header