ab_core_primitives::block::header

Module owned

Data structures related to the owned version of BlockHeader

Structs

OwnedBeaconChainBlockHeader

An owned version of BeaconChainBlockHeader.

OwnedBeaconChainBlockHeaderUnsealed

Owned beacon chain block header, which is not sealed yet

OwnedIntermediateShardBlockHeader

An owned version of IntermediateShardBlockHeader.

OwnedIntermediateShardBlockHeaderUnsealed

Owned intermediate shard block header, which is not sealed yet

OwnedLeafShardBlockHeader

An owned version of LeafShardBlockHeader.

OwnedLeafShardBlockHeaderUnsealed

Owned leaf shard block header, which is not sealed yet

Enums

OwnedBeaconChainBlockHeaderError

Errors for OwnedBeaconChainBlockHeader

OwnedBlockHeader

An owned version of BlockHeader.

OwnedBlockHeaderError

Errors for OwnedBlockHeader

OwnedIntermediateShardBlockHeaderError

Errors for OwnedIntermediateShardBlockHeader